[#2]환경변수 치환 및 종료코드 처리

+++ minishell.c
@@ -6,7 +6,7 @@
 /*   By: siychoi <siychoi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/01 17:24:50 by siychoi           #+#    #+#             */
-/*   Updated: 2024/06/09 12:30:30 by siychoi          ###   ########.fr       */
+/*   Updated: 2024/06/12 13:20:38 by siychoi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,12 +48,12 @@ void	exec_minishell(t_envp **my_envp)
 		{
 			printf("\033[u\033[1B\033[1Aexit\n");
 			tcsetattr(STDIN_FILENO, TCSANOW, &old_term);
-			exit(1);
+			exit(0);
 		}
 		else if (ft_strlen(buffer) != 0)
 		{
 			list = lexer(buffer);
-		 	add_history(buffer);
+			add_history(buffer);
 			redirection_to_filename(list, &flag);
 			change_dollar_all_tokens(list, my_envp);
 			if (flag == TRUE)
@@ -74,31 +74,27 @@ void	exec_minishell(t_envp **my_envp)
 
 int	exe_one_command(t_envp **my_envp, t_phrase *phrase)
 {
-	int	infile_fd;
-	int	outfile_fd;
-	int	stdin_fd;
-	int	stdout_fd;
+	int	fd[4];
 	int	return_code;
-	
-	infile_fd = 0;
-	outfile_fd = 1;
-	if (phrase->head == NULL)
+
+	fd[0] = 0;
+	fd[1] = 1;
+	if (phrase->head == NULL && phrase->infile_type == 3)
 	{
-		if (phrase->infile_type == 3)
-			unlink(phrase->infile_name);
+		unlink(phrase->infile_name);
 		return (0);
 	}
 	if (is_builtin_cmd(phrase))
 	{
-		stdin_fd = dup(0);
-		stdout_fd = dup(1);
-		if (open_in_and_out_fd(phrase, &infile_fd, &outfile_fd) != TRUE)
+		fd[2] = dup(0);
+		fd[3] = dup(1);
+		if (open_in_and_out_fd(phrase, &fd[0], &fd[1]) != TRUE)
 			return (errno);
 		return_code = exe_only_builtin_cmd(my_envp, phrase);
 		if (phrase->infile_type == 3)
 			unlink(phrase->infile_name);
-		dup2(stdin_fd, 0);
-		dup2(stdout_fd, 1);
+		dup2(fd[2], 0);
+		dup2(fd[3], 1);
 		return (return_code);
 	}
 	else
@@ -129,7 +125,7 @@ int	wait_and_return(t_input *list, int last_code)
 	return (code);
 }
 
-char	**token_to_arr(t_phrase *phrase) 
+char	**token_to_arr(t_phrase *phrase)
 {
 	char	**result;
 	int		i;
@@ -163,11 +159,17 @@ int	exe_only_builtout_cmd(t_envp **my_envp, t_phrase *phrase)
 	pid_t	pid;
 	char	**envp;
 	int		status;
+	int		infile_fd;
+	int		outfile_fd;
 
 	pid = fork();
 	if (pid == 0)
 	{
 		set_child_signal();
+		infile_fd = 0;
+		outfile_fd = 1;
+		if (open_in_and_out_fd(phrase, &infile_fd, &outfile_fd) != TRUE)
+			exit(errno);
 		envp = envp_list_to_arr(my_envp);
 		return (child_process_exe(phrase, envp));
 	}
